test(integration): move controller batch 2
